requirements.txt
requests
pandas
aiohttp
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
azure-identity